src/backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Django==3.2
django-filter==22.1
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.0
drf-yasg==1.21.4
drf-yasg2==1.19.4
gunicorn==19.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.5
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
